Update user schema
Only skills reside in DynamoDB now, the rest are retrieved from
Cognito.

For User only the option to retrieve all are present, due to how
the requests are made and they will all be needed at the same time.
The details for an individual user can be retrieved from their own
token.

Cognito has been removed from serverless as needs to remain persistent
due to the custom domain.